requirements.txt
streamlit
pdfplumber
python-docx
nltk
pandas
scikit-learn
